requirements.txt
fastapi
uvicorn
tortoise-orm
passlib
python-jose
pydantic
aerich
bcrypt
python-dotenv
